requirements.txt
dataclasses==0.6
flake8==4.0.1
mccabe==0.6.1
neo4j==4.4.3
numpy==1.22.4
pandas==1.4.2
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
